Extract the incoming verification requests details view to its own struct
